fix usher url, add timeout
so I'm not sure when this happened, but twitch switched from using `usher.twitch.tv` to `usher.ttvnw.net`. It's possible it was redirecting (301/302), and simply going on unannounced till now when the old url just infinitely hangs. Thankfully this was a very simple fix, and a hotfix release will need to be released soon. I should also add timeouts for all the requests along with proper handling of those timeouts.